chore: move implementation of bitwise operations into `blackbox_solver`
